// src/constants/enums/register.ts
export enum RegisterForm {
  id = "register__form",

  autocomplete_inputs = "off",
  input_size = "medium",
  button_size = "large",

  firstName_input_id = "firstName_input_id",
  firstName_input_name = "firstName_input",
  firstName_input_label = "",
  firstName_input_role = "firstName_input",
  firstName_input_arial_label = "first-name-input",
  firstName_input_placeholder = "Your first name",
  firstName_input_type = "text",

  lastName_input_id = "lastName_input_id",
  lastName_input_name = "lastName_input",
  lastName_input_label = "",
  lastName_input_role = "lastName_input",
  lastName_input_arial_label = "last-name-input",
  lastName_input_placeholder = "Your last name",
  lastName_input_type = "text",

  email_input_id = "email_input_id",
  email_input_name = "email_input",
  email_input_label = "",
  email_input_role = "email-input",
  email_input_arial_label = "email-input",
  email_input_placeholder = "Your email address",
  email_input_type = "email",

  password_input_id = "password_input_id",
  password_input_name = "password_input",
  password_input_label = "",
  password_input_role = "password-input",
  password_input_arial_label = "password-input",
  password_input_placeholder = "Your password",
  password_input_type_hidden = "password",
  password_input_type_show = "text",

  confirmPassword_input_id = "confirmPassword_input_id",
  confirmPassword_input_name = "confirmPassword_input",
  confirmPassword_input_label = "",
  confirmPassword_input_role = "confirmPassword-input",
  confirmPassword_input_arial_label = "confirmPassword-input",
  confirmPassword_input_placeholder = "Confirm password",
  confirmPassword_input_type_hidden = "password",
  confirmPassword_input_type_show = "text",
}

export enum RegisterFormTexts {
  form_title = "Create an account",
  registered_user = "already have an account?",
  register_button = "Join now",
  login_button = "Login"
}